Added word lookup and thesaurus orginization.
